feat: [prompt-hinter-004] - [Prove prompt hinting and validation behavior end to end]
