+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -74,6 +74,7 @@ work inputs.
 - When current-selection workstation editing needs canonical editable factory data, enable the editable definition query whenever a workstation node is selected, including initial widget mount; selection-transition-only gates leave already-selected workstation editors stuck without worker or prompt values.
 - When current-selection workstation editing needs prompt-variable help, keep the query boundary in `ui/src/features/current-selection/useCurrentWorkstationPromptTemplateContract.ts`, map its result into explicit `loading`, `empty`, `error`, and `ready` form state inside `use-editable-workstation-configuration-state.ts`, and render the disclosure in `workstation-editable-configuration-section.tsx`; that keeps fetch logic out of the field renderer while preserving one typed help contract per selected workstation.
 - When current-selection workstation editing needs inline prompt diagnostics, keep authoritative validation in `useCurrentWorkstationPromptTemplateValidation.ts`, fold its `idle`/`loading`/`error`/`ready` states into `use-editable-workstation-configuration-state.ts`, and render the textarea overlay plus diagnostic list from `workstation-prompt-field.tsx`; that keeps save-blocking, inline squiggles, and supporting copy driven by one validation contract instead of component-local parsing.
+- When current-selection tests mount `CurrentSelectionWidget` with the real prompt-help or prompt-validation hooks, render through a `QueryClientProvider` wrapper and keep App-shell fetch assertions scoped to the endpoint under test; selected-workstation shells now issue prompt-template contract queries during normal editor-state setup.
 - When a live factory-definition save must refresh already-open dashboards, emit a canonical `FACTORY_CHANGE` event with the same factory payload shape used by `INITIAL_STRUCTURE_REQUEST`, then update both replay reducers (`pkg/factory/projections/world_state.go`, `ui/src/features/timeline/state/timeline/replayWorldState.ts`) and the editable current-factory cache bridge in `ui/src/features/dashboard/useDashboardSnapshot.ts` together.
 - When UI replay or topology route families change, update all three dashboard seams together: `ui/src/features/timeline/state/timeline/replayFactoryTopology.ts`, `ui/src/features/timeline/state/timeline/projectTopology.ts`, and `ui/src/api/dashboard/types.ts`; replay placement, graph edges, and typed outcome kinds otherwise drift apart.
 - When workstation-request current-selection moves inference-owned detail between dispatch summaries and inference attempts, update all four seams together: the authored factory-world projection schemas under `api/components/schemas/factory-world/`, regenerated Go and TypeScript API artifacts, handwritten dashboard request adapters in `ui/src/features/current-selection/` and `ui/src/features/timeline/state/timeline/`, and the selected-work current-selection tests that prove prompts and responses render from attempts rather than dispatch summaries.
