+++ ui/scripts/verify-import-export-storybook-responsive.test.mjs
@@ -1,11 +1,9 @@
 import { describe, expect, test, vi } from "vitest";
-
 import {
   expectDialogWithinViewport,
   expectNoHorizontalOverflow,
   expectOrderedLeftEdges,
   expectVisible,
-  verifyDashboardHeader,
   verifyExportDialog,
   verifyImportDialog,
   verifyStory,
@@ -13,16 +11,13 @@ import {
   waitForStoryRegion,
   waitForStoryRender,
 } from "./verify-import-export-storybook-responsive.mjs";
-
 describe("waitForStoryRender", () => {
   test("waits for the Storybook root selector and rendered children", async () => {
     const page = {
       waitForFunction: vi.fn().mockResolvedValue(undefined),
       waitForSelector: vi.fn().mockResolvedValue(undefined),
     };
-
     await waitForStoryRender(page);
-
     expect(page.waitForSelector).toHaveBeenCalledWith("#storybook-root", {
       state: "attached",
       timeout: 30000,
@@ -38,9 +33,7 @@ describe("waitForStoryRender", () => {
       waitForFunction: vi.fn().mockResolvedValue(undefined),
       waitForSelector: vi.fn().mockResolvedValue(undefined),
     };
-
     await waitForStoryRender(page);
-
     const readyCheck = page.waitForFunction.mock.calls[0]?.[0];
     if (typeof readyCheck !== "function") {
       throw new Error("expected Storybook readiness predicate");
@@ -124,7 +117,6 @@ describe("verifyStory", () => {
       newPage: vi.fn().mockResolvedValue(page),
     };
     const failure = new Error("story failed");
-
     await expect(
       verifyStory(
         browser,
@@ -149,9 +141,7 @@ describe("story locators", () => {
     const page = {
       getByRole: vi.fn().mockReturnValue(dialog),
     };
-
     const result = await waitForDialog(page, "Export factory");
-
     expect(page.getByRole).toHaveBeenCalledWith("dialog", {
       name: "Export factory",
     });
@@ -166,9 +156,7 @@ describe("story locators", () => {
     const page = {
       getByRole: vi.fn().mockReturnValue(region),
     };
-
     const result = await waitForStoryRegion(page, "dashboard summary");
-
     expect(page.getByRole).toHaveBeenCalledWith("region", {
       name: "dashboard summary",
     });
@@ -380,122 +368,5 @@ describe("story assertions", () => {
     expect(dialog.getByRole).toHaveBeenCalledWith("button", {
       name: "Activate factory",
     });
-  });
-});
-
-describe("verifyDashboardHeader", () => {
-  test("verifyDashboardHeader exercises keyboard and desktop ordering checks", async () => {
-    const heading = {
-      boundingBox: vi.fn().mockResolvedValue({ height: 20, width: 120, x: 10, y: 0 }),
-      getByText: vi.fn().mockReturnValue({
-        getAttribute: vi.fn().mockResolvedValue("sr-only"),
-        isVisible: vi.fn().mockResolvedValue(true),
-      }),
-      isVisible: vi.fn().mockResolvedValue(true),
-    };
-    const slider = {
-      boundingBox: vi.fn().mockResolvedValue({ height: 20, width: 200, x: 160, y: 0 }),
-      focus: vi.fn().mockResolvedValue(undefined),
-      isVisible: vi.fn().mockResolvedValue(true),
-    };
-    const streamStatus = {
-      boundingBox: vi.fn().mockResolvedValue({ height: 20, width: 180, x: 380, y: 0 }),
-      isVisible: vi.fn().mockResolvedValue(true),
-    };
-    const currentButton = {
-      focus: vi.fn().mockResolvedValue(undefined),
-      isVisible: vi.fn().mockResolvedValue(true),
-    };
-    const exportButton = {
-      boundingBox: vi.fn().mockResolvedValue({ height: 20, width: 120, x: 580, y: 0 }),
-      isVisible: vi.fn().mockResolvedValue(true),
-    };
-    const currentTick = { isVisible: vi.fn().mockResolvedValue(true) };
-    const toolbar = {
-      getByRole: vi.fn((role, options) => {
-        if (role === "heading") {
-          return heading;
-        }
-        if (role === "slider") {
-          return slider;
-        }
-        if (role === "status") {
-          return streamStatus;
-        }
-        if (options?.name === "Return to current tick") {
-          return currentButton;
-        }
-        return exportButton;
-      }),
-      waitFor: vi.fn().mockResolvedValue(undefined),
-    };
-    const page = {
-      evaluate: vi.fn().mockResolvedValue({ clientWidth: 1440, scrollWidth: 1440 }),
-      getByRole: vi.fn().mockReturnValue(toolbar),
-      getByText: vi
-        .fn()
-        .mockReturnValueOnce(currentTick)
-        .mockReturnValueOnce({ isVisible: vi.fn().mockResolvedValue(true) })
-        .mockReturnValueOnce(currentTick),
-      keyboard: {
-        press: vi.fn().mockResolvedValue(undefined),
-      },
-    };
-
-    await verifyDashboardHeader(page, null, {
-      height: 900,
-      label: "desktop",
-      width: 1440,
-    });
-
-    expect(page.getByRole).toHaveBeenCalledWith("region", {
-      name: "dashboard summary",
-    });
-    expect(slider.focus).toHaveBeenCalledTimes(1);
-    expect(currentButton.focus).toHaveBeenCalledTimes(1);
-    expect(page.keyboard.press).toHaveBeenNthCalledWith(1, "ArrowLeft");
-    expect(page.keyboard.press).toHaveBeenNthCalledWith(2, "Enter");
-  });
-
-  test("verifyDashboardHeader rejects a non-sr-only heading wordmark", async () => {
-    const toolbar = {
-      getByRole: vi.fn((role, _options) => {
-        if (role === "heading") {
-          return {
-            getByText: vi.fn().mockReturnValue({
-              getAttribute: vi.fn().mockResolvedValue("text-visible"),
-              isVisible: vi.fn().mockResolvedValue(true),
-            }),
-            isVisible: vi.fn().mockResolvedValue(true),
-          };
-        }
-        if (role === "slider") {
-          return { focus: vi.fn().mockResolvedValue(undefined), isVisible: vi.fn().mockResolvedValue(true) };
-        }
-        if (role === "status") {
-          return { isVisible: vi.fn().mockResolvedValue(true) };
-        }
-        return { isVisible: vi.fn().mockResolvedValue(true) };
-      }),
-      waitFor: vi.fn().mockResolvedValue(undefined),
-    };
-    const page = {
-      evaluate: vi.fn().mockResolvedValue({ clientWidth: 390, scrollWidth: 390 }),
-      getByRole: vi.fn().mockReturnValue(toolbar),
-      getByText: vi.fn().mockReturnValue({ isVisible: vi.fn().mockResolvedValue(true) }),
-      keyboard: {
-        press: vi.fn().mockResolvedValue(undefined),
-      },
-    };
-
-    await expect(
-      verifyDashboardHeader(page, null, {
-        height: 844,
-        label: "mobile",
-        width: 390,
-      }),
-    ).rejects.toThrow(
-      "Dashboard heading wordmark was not hidden with sr-only styling.",
-    );
   });
 });
