+++ ui/src/features/current-selection/current-selection-widget.test.tsx
@@ -284,7 +284,7 @@ describe("CurrentSelectionWidget", () => {
       workItem,
     } = buildSelectedWorkItemFixture();
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectedNode,
@@ -652,7 +652,7 @@ describe("CurrentSelectionWidget", () => {
       );
     }
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectedStatePlace,
@@ -697,7 +697,7 @@ describe("CurrentSelectionWidget", () => {
       );
     }
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectStateWorkItem,
@@ -742,7 +742,7 @@ describe("CurrentSelectionWidget", () => {
         attempt.workstation_name === selectedNode.workstation_name,
     );
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectedNode,
@@ -769,7 +769,7 @@ describe("CurrentSelectionWidget", () => {
   });
 
   it("does not load the editable factory definition when no workstation is selected", () => {
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -785,7 +785,7 @@ describe("CurrentSelectionWidget", () => {
   it("enables editable workstation loading after a workstation becomes selected", () => {
     const snapshot = semanticWorkflowDashboardSnapshot;
     const selectedNode = snapshot.topology.workstation_nodes_by_id.review;
-    const { rerender } = render(
+    const { rerender } = renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -816,7 +816,7 @@ describe("CurrentSelectionWidget", () => {
       buildEditableDefinitionResult(buildEditableFactoryDefinition()),
     );
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectedNode,
@@ -848,7 +848,7 @@ describe("CurrentSelectionWidget", () => {
       buildEditableDefinitionResult(buildEditableFactoryDefinition()),
     );
 
-    const { rerender } = render(
+    const { rerender } = renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -892,7 +892,7 @@ describe("CurrentSelectionWidget", () => {
       buildEditableDefinitionResult(buildEditableFactoryDefinition()),
     );
 
-    const { rerender } = render(
+    const { rerender } = renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -980,7 +980,7 @@ describe("CurrentSelectionWidget", () => {
       },
     );
 
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection({
           selectedNode:
@@ -1036,7 +1036,7 @@ describe("CurrentSelectionWidget", () => {
   });
 
   it("renders the empty current-selection guidance when nothing is selected", () => {
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -1052,7 +1052,7 @@ describe("CurrentSelectionWidget", () => {
   });
 
   it("renders localized current-selection shell copy for a supported non-default locale", () => {
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         locale="ja"
@@ -1076,7 +1076,7 @@ describe("CurrentSelectionWidget", () => {
   });
 
   it("renders disabled undo and redo controls in the shared current-selection header by default", () => {
-    render(
+    renderWithQueryClient(
       <CurrentSelectionWidget
         currentSelection={buildCurrentSelection()}
         now={DETAIL_CARD_NOW}
@@ -1133,9 +1133,11 @@ function renderWithQueryClient(view: ReactNode) {
     },
   });
 
-  return render(
-    <QueryClientProvider client={queryClient}>{view}</QueryClientProvider>,
-  );
+  return render(view, {
+    wrapper: ({ children }) => (
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    ),
+  });
 }
 
 function expandEditableConfiguration() {
